requirements.txt
Django==3.1.7
djangorestframework==3.12.2
dj-database-url==0.2.1
psycopg2-binary==2.8.2
flake8==3.8.4
black==20.8b1
coverage==5.4